cambie la vista de las card de producto en el inicio para la vista de celulares y cambie de lugar la logica de agregar al carrito
