+++ src/components/pages/DetalleProducto.jsx
@@ -1,7 +1,8 @@
 import { Card, Row, Col, Breadcrumb } from "react-bootstrap";
 import { Link, useParams } from "react-router";
-import { obtenerProductosIdAPI, agregarAlCarrito } from "../../helpers/queries.js";
+import { obtenerProductosIdAPI } from "../../helpers/queries.js";
 import { useEffect, useState } from "react";
+import Swal from "sweetalert2";
 
 const DetalleProducto = () => {
   const [producto, setProducto] = useState(null);
@@ -23,6 +24,39 @@ const DetalleProducto = () => {
     } else {
       alert("Ocurrió un error");
     }
+  };
+
+  const agregarAlCarrito = (producto, cantidadDeseada = 1) => {
+    const usuarioLogueado = JSON.parse(sessionStorage.getItem("usuarioKey"));
+
+    if (!usuarioLogueado) {
+      Swal.fire({
+        icon: "warning",
+        title: "Atención",
+        text: "Debes iniciar sesión para realizar esta acción",
+        confirmButtonText: "Entendido",
+        confirmButtonColor: "#f0ad4e",
+      });
+      return;
+    }
+const carritoActual = JSON.parse(localStorage.getItem("carrito")) || [];
+    const indice = carritoActual.findIndex((item) => item._id === producto._id);
+    
+    if (indice !== -1) {
+      carritoActual[indice].cantidad += cantidadDeseada;
+    } else {
+      carritoActual.push({ ...producto, cantidad: cantidadDeseada });
+    }
+    
+    localStorage.setItem("carrito", JSON.stringify(carritoActual));
+    
+    Swal.fire({
+      icon: "success",
+      title: "Producto agregado",
+      text: "El producto fue agregado al carrito correctamente.",
+      timer: 1500,
+      showConfirmButton: false,
+    });
   };
 
   if (!producto) {
