+++ src/helpers/queries.js
@@ -280,36 +280,7 @@ export const cambiarEstadoPedidoAPI = async (id, estado) => {
   }
 };
 
-export const agregarAlCarrito = (producto, cantidadDeseada = 1) => {
-  const usuarioLogueado = JSON.parse(sessionStorage.getItem("usuarioKey"));
 
-  if (!usuarioLogueado) {
-    Swal.fire({
-      icon: "warning",
-      title: "Atención",
-      text: "Debes iniciar sesión para realizar esta acción",
-      confirmButtonText: "Entendido",
-      confirmButtonColor: "#f0ad4e",
-    });
-    return;
-  }
-
-  const carritoActual = JSON.parse(localStorage.getItem("carrito")) || [];
-  const indice = carritoActual.findIndex((item) => item._id === producto._id);
-  if (indice !== -1) {
-    carritoActual[indice].cantidad += cantidadDeseada;
-  } else {
-    carritoActual.push({ ...producto, cantidad: cantidadDeseada });
-  }
-  localStorage.setItem("carrito", JSON.stringify(carritoActual));
-  Swal.fire({
-    icon: "success",
-    title: "Producto agregado",
-    text: "El producto fue agregado al carrito correctamente.",
-    timer: 1500,
-    showConfirmButton: false,
-  });
-};
 
 export const crearPedidoAPI = async (pedido) => {
   try {
